ES6 modulize core/utils
