+++ cocos/core/utils/binary-search.js
@@ -83,7 +83,7 @@ var EPSILON = 1e-6;
 //     return ~l;
 // }
 
-function binarySearchEpsilon (array, value) {
+export default function binarySearchEpsilon (array, value) {
     for (var l = 0, h = array.length - 1, m = h >>> 1;
          l <= h;
          m = (l + h) >>> 1
@@ -101,8 +101,3 @@ function binarySearchEpsilon (array, value) {
     }
     return ~l;
 }
-
-
-module.exports = {
-    binarySearchEpsilon
-};
