+++ cocos/core/utils/decode-uuid.js
@@ -24,7 +24,7 @@
  THE SOFTWARE.
  ****************************************************************************/
 
-var Base64Values = require('./misc').BASE64_VALUES;
+import {BASE64_VALUES} from 'misc';
 
 var HexChars = '0123456789abcdef'.split('');
 
@@ -33,15 +33,15 @@ var UuidTemplate = _t.concat(_t, '-', _t, '-', _t, '-', _t, '-', _t, _t, _t);
 var Indices = UuidTemplate.map(function (x, i) { return x === '-' ? NaN : i; }).filter(isFinite);
 
 // fcmR3XADNLgJ1ByKhqcC5Z -> fc991dd7-0033-4b80-9d41-c8a86a702e59
-module.exports = function (base64) {
+export default function decodeUuid (base64) {
     if (base64.length !== 22) {
         return base64;
     }
     UuidTemplate[0] = base64[0];
     UuidTemplate[1] = base64[1];
     for (var i = 2, j = 2; i < 22; i += 2) {
-        var lhs = Base64Values[base64.charCodeAt(i)];
-        var rhs = Base64Values[base64.charCodeAt(i + 1)];
+        var lhs = BASE64_VALUES[base64.charCodeAt(i)];
+        var rhs = BASE64_VALUES[base64.charCodeAt(i + 1)];
         UuidTemplate[Indices[j++]] = HexChars[lhs >> 2];
         UuidTemplate[Indices[j++]] = HexChars[((lhs & 3) << 2) | rhs >> 4];
         UuidTemplate[Indices[j++]] = HexChars[rhs & 0xF];
@@ -50,5 +50,5 @@ module.exports = function (base64) {
 };
 
 if (CC_TEST) {
-    cc._Test.decodeUuid = module.exports;
+    cc._Test.decodeUuid = decodeUuid;
 }
